Min throughput workaround implemented.
Min throughput of a Cosmos database can be changed after the database is scaled out. In this case Bullfrog might not be able to reset the throughput to its lowest level. The workaround here uses the error message to find out the current valid min throughput. The cosmos client library does not have support for getting this value currently.